Record the real output in the distribution docstring examples
pytest --doctest-modules over distributions/ is now green: 139 examples,
no failures, down from 39 failing.

Most were the numpy 2 scalar repr. Weibull.mean(3, 4) prints
np.float64(2.7192074311664314); the docstring recorded the bare float
numpy 1 used to print. The examples now carry the wrapper because that
is what a reader sees at their own prompt. The alternative was
np.set_printoptions(legacy="1.25") in a fixture, which keeps the
docstrings prettier by showing people output their session will not
produce — prettier, but not true.

Four qf examples exceeded 79 columns once the real output was recorded,
numpy having rewrapped the arrays differently from the hand-wrapping.
Rather than reflow them into something numpy would never emit, those
examples now take fewer probabilities, so what is printed is exactly
what that input produces.

Two scalar examples had drifted in the last digit and are re-recorded
against a direct run.

Every rewrite was gated on the numbers agreeing to 1e-12 first, so a
genuine mismatch could not be blessed by the sweep; nothing failed that
gate. Two files needed hand correction afterwards where mean and moment
share a value and the text search matched the wrong occurrence.

+++ surpyval/univariate/parametric/distributions/binomial.py
@@ -64,7 +64,7 @@ class Binomial_(DiscreteParametricFitter):
         --------
         >>> from surpyval import Binomial
         >>> Binomial.df(2, 5, 0.3)
-        0.3086999999999998
+        np.float64(0.3086999999999998)
         """
         return binom.pmf(x, n, p)
 
@@ -97,7 +97,7 @@ class Binomial_(DiscreteParametricFitter):
         --------
         >>> from surpyval import Binomial
         >>> Binomial.ff(2, 5, 0.3)
-        0.83692
+        np.float64(0.83692)
         """
         return binom.cdf(x, n, p)
 
@@ -129,7 +129,7 @@ class Binomial_(DiscreteParametricFitter):
         --------
         >>> from surpyval import Binomial
         >>> Binomial.sf(2, 5, 0.3)
-        0.16308
+        np.float64(0.16308)
         """
         return binom.sf(x, n, p)
 
@@ -215,7 +215,7 @@ class Binomial_(DiscreteParametricFitter):
         --------
         >>> from surpyval import Binomial
         >>> Binomial.qf(0.5, 5, 0.3)
-        1.0
+        np.float64(1.0)
         """
         return binom.ppf(q, n, p)
 
@@ -289,7 +289,7 @@ class Binomial_(DiscreteParametricFitter):
         --------
         >>> from surpyval import Binomial
         >>> Binomial.moment(1, 5, 0.3)
-        1.5
+        np.float64(1.5)
         """
         return binom.moment(m, n, p)
 
@@ -302,7 +302,7 @@ class Binomial_(DiscreteParametricFitter):
         --------
         >>> from surpyval import Binomial
         >>> Binomial.entropy(5, 0.3)
-        1.413614855283445
+        np.float64(1.413614855283445)
         """
         return binom.entropy(n, p)
 
@@ -414,7 +414,7 @@ class Binomial_(DiscreteParametricFitter):
         >>> from surpyval import Binomial
         >>> model = Binomial.from_params([5, 0.3])
         >>> model.mean()
-        1.5
+        np.float64(1.5)
         """
         params = np.atleast_1d(np.asarray(params, dtype=float))
